Rearrange modules to satisfy clippy.
